petsy1/backend/requirements.txt
# ==== Core Framework ====
Flask==3.0.3

# ==== Database ====
mysql-connector-python==9.0.0

# ==== Email / OTP ====
Flask-Mail==0.10.0

# ==== CORS (frontend-backend connection) ====
Flask-Cors==4.0.0

# ==== Date & Time Utilities ====
pytz==2024.1

# ==== Security / Passwords ====
Werkzeug==3.0.2

# ==== Deployment (for production, optional) ====
gunicorn==23.0.0

Pillow==10.3.0

PyJWT==2.9.0

python-dotenv==1.0.1